Substituição de todos os valores do tipo double para BigDecimal

+++ MundoDoEstudante/src/br/com/MundoDoEstudante/classes/Gratificacao.java
@@ -1,10 +1,12 @@
 package br.com.MundoDoEstudante.classes;
 
+import java.math.BigDecimal;
+
 public interface Gratificacao {
 	
 
-	double getGraficacao();
-	void setGratificacaoSemana(double valor);
+	BigDecimal getGraficacao();
+	void setGratificacaoSemana(BigDecimal valor);
 	
 
 }
