+++ MundoDoEstudante/src/br/com/MundoDoEstudante/classes/calculadoraGratificacao/LapisCalculadoraGratificacao.java
@@ -1,5 +1,6 @@
 package br.com.MundoDoEstudante.classes.calculadoraGratificacao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import br.com.MundoDoEstudante.classes.Atendente;
@@ -11,9 +12,9 @@ public class LapisCalculadoraGratificacao extends CalculadorGratificacao {
 	private static final double TERCEIRO_COLOCADO = 0.006;
 	private static final double DEMAIS_COLOCADOS = 0.005;
 
-
 	public List<Atendente> calcularGratificacao() {
-		return super.calcularGratificacao(PRIMEIRO_COLOCADO, SEGUNDO_COLOCADO, TERCEIRO_COLOCADO, DEMAIS_COLOCADOS);
+		return super.calcularGratificacao(BigDecimal.valueOf(PRIMEIRO_COLOCADO), BigDecimal.valueOf(SEGUNDO_COLOCADO),
+				BigDecimal.valueOf(TERCEIRO_COLOCADO), BigDecimal.valueOf(DEMAIS_COLOCADOS));
 	}
 
 }
